+++ src/components/link/link.component.tsx
@@ -47,6 +47,8 @@ export interface LinkProps extends StyledLinkProps, React.AriaAttributes {
   ariaLabel?: string;
   /** allows to set rel property in <a> tag */
   rel?: string;
+  /** @ignore @private internal prop to be set when no href or onClick passed */
+  placeholderTabIndex?: boolean;
 }
 
 export const Link = React.forwardRef<
@@ -72,6 +74,7 @@ export const Link = React.forwardRef<
       target,
       variant = "default",
       isDarkBackground,
+      placeholderTabIndex,
       ...rest
     }: LinkProps,
     ref
@@ -136,6 +139,7 @@ export const Link = React.forwardRef<
       "aria-label": ariaLabel,
       ...ariaProps,
     };
+
     const createLinkBasedOnType = () => {
       let type = "a";
 
@@ -145,7 +149,12 @@ export const Link = React.forwardRef<
 
       return React.createElement(
         type,
-        componentProps,
+        {
+          ...componentProps,
+          ...(placeholderTabIndex &&
+            href === undefined &&
+            !onClick && { tabIndex: -1 }),
+        },
         <>
           {renderLinkIcon()}