+++ src/components/menu/menu.context.js
@@ -3,6 +3,6 @@ import React from "react";
 export default React.createContext({
   menuType: "light",
   inMenu: false,
-  openSubmenuIndex: null,
-  setOpenSubmenuIndex: /* istanbul ignore next */ () => {},
+  openSubmenuId: null,
+  setOpenSubmenuId: /* istanbul ignore next */ () => {},
 });