+++ src/components/select/filterable-select/filterable-select.component.js
@@ -56,6 +56,8 @@ const FilterableSelect = React.forwardRef(
       listPlacement = "bottom",
       flipEnabled = true,
       inputRef,
+      enableVirtualScroll,
+      virtualScrollOverscan,
       ...textboxProps
     },
     ref
@@ -539,6 +541,8 @@ const FilterableSelect = React.forwardRef(
         listPlacement={listPlacement}
         flipEnabled={flipEnabled}
         isOpen={isOpen}
+        enableVirtualScroll={enableVirtualScroll}
+        virtualScrollOverscan={virtualScrollOverscan}
       >
         {children}
       </FilterableSelectList>
@@ -618,6 +622,13 @@ FilterableSelect.propTypes = {
   listPlacement: PropTypes.oneOf(["top", "bottom", "right", "left"]),
   /** Use the opposite list placement if the set placement does not fit */
   flipEnabled: PropTypes.bool,
+  /** Set this prop to enable a virtualised list of options. If it is not used then all options will be in the
+   * DOM at all times, which may cause performance problems on very large lists */
+  enableVirtualScroll: PropTypes.bool,
+  /** The number of options to render into the DOM at once, either side of the currently-visible ones.
+   * Higher values make for smoother scrolling but may impact performance.
+   * Only used if the `enableVirtualScroll` prop is set. */
+  virtualScrollOverscan: PropTypes.number,
 };
 
 FilterableSelect.defaultProps = {