+++ src/components/select/multi-select/multi-select.component.js
@@ -67,6 +67,8 @@ const MultiSelect = React.forwardRef(
       flipEnabled = true,
       wrapPillText = true,
       inputRef,
+      enableVirtualScroll,
+      virtualScrollOverscan,
       ...textboxProps
     },
     ref
@@ -442,10 +444,6 @@ const MultiSelect = React.forwardRef(
         const isAlreadySelected =
           actualValue.findIndex((val) => isExpectedValue(val, newValue)) !== -1;
 
-        if (!isAlreadySelected && isControlled.current && onChange) {
-          onChange(createCustomEvent([...actualValue, newValue]));
-        }
-
         textboxRef.focus();
         isMouseDownReported.current = false;
 
@@ -457,7 +455,7 @@ const MultiSelect = React.forwardRef(
           return [...previousValue, newValue];
         });
       },
-      [createCustomEvent, onChange, textboxRef, actualValue, updateValue]
+      [textboxRef, actualValue, updateValue]
     );
 
     function onSelectListClose() {
@@ -540,6 +538,8 @@ const MultiSelect = React.forwardRef(
         loaderDataRole="multi-select-list-loader"
         multiselectValues={actualValue}
         isOpen={isOpen}
+        enableVirtualScroll={enableVirtualScroll}
+        virtualScrollOverscan={virtualScrollOverscan}
       >
         {children}
       </FilterableSelectList>
@@ -629,6 +629,13 @@ MultiSelect.propTypes = {
   flipEnabled: PropTypes.bool,
   /** Wraps the pill text when it would overflow the input width */
   wrapPillText: PropTypes.bool,
+  /** Set this prop to enable a virtualised list of options. If it is not used then all options will be in the
+   * DOM at all times, which may cause performance problems on very large lists */
+  enableVirtualScroll: PropTypes.bool,
+  /** The number of options to render into the DOM at once, either side of the currently-visible ones.
+   * Higher values make for smoother scrolling but may impact performance.
+   * Only used if the `enableVirtualScroll` prop is set. */
+  virtualScrollOverscan: PropTypes.number,
 };
 
 MultiSelect.defaultProps = {