+++ src/components/select/multi-select/multi-select.test.js
@@ -14,8 +14,9 @@ import {
 } from "../../../../cypress/locators";
 
 import {
-  selectList,
+  selectListWrapper,
   selectOption,
+  selectOptionByText,
   dropdownButton,
   selectListText,
   multiColumnsSelectListHeader,
@@ -43,7 +44,7 @@ import {
   verifyRequiredAsteriskForLabel,
 } from "../../../../cypress/support/component-helper/common-steps";
 
-import { keyCode, positionOfElement } from "../../../../cypress/support/helper";
+import { positionOfElement } from "../../../../cypress/support/helper";
 import {
   SIZE,
   CHARACTERS,
@@ -470,6 +471,23 @@ const MultiSelectCustomColorComponent = ({ ...props }) => {
   );
 };
 
+const MultiSelectWithManyOptionsAndVirtualScrolling = () => (
+  <MultiSelect
+    name="virtualised"
+    id="virtualised"
+    label="choose an option"
+    labelInline
+    enableVirtualScroll
+    virtualScrollOverscan={10}
+  >
+    {Array(10000)
+      .fill()
+      .map((_, index) => (
+        <Option key={index} value={`${index}`} text={`Option ${index + 1}.`} />
+      ))}
+  </MultiSelect>
+);
+
 const testData = [CHARACTERS.DIACRITICS, CHARACTERS.SPECIALCHARACTERS];
 const testPropValue = CHARACTERS.STANDARD;
 const columns = 3;
@@ -600,7 +618,7 @@ context("Tests for Multi Select component", () => {
 
       commonDataElementInputPreview().should("have.attr", "readOnly");
       selectInput().click();
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it.each([
@@ -737,7 +755,7 @@ context("Tests for Multi Select component", () => {
       commonDataElementInputPreview()
         .should("be.focused")
         .and("have.attr", "aria-expanded", "false");
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it("should not open the list with mouse click on Multi Select input", () => {
@@ -747,14 +765,14 @@ context("Tests for Multi Select component", () => {
       commonDataElementInputPreview()
         .should("be.focused")
         .and("have.attr", "aria-expanded", "false");
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it("should open the list with mouse click on dropdown button", () => {
       CypressMountWithProviders(<MultiSelectComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
     });
 
     it("should close the list with mouse click on dropdown button", () => {
@@ -762,42 +780,42 @@ context("Tests for Multi Select component", () => {
 
       dropdownButton().click();
       dropdownButton().click();
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it("should close the list with the Tab key", () => {
       CypressMountWithProviders(<MultiSelectComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
-      selectInput().tab();
+      selectListWrapper().should("be.visible");
+      selectInput().realPress("Tab");
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it("should close the list with the Esc key", () => {
       CypressMountWithProviders(<MultiSelectComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
-      selectInput().trigger("keydown", { ...keyCode("Esc") });
+      selectListWrapper().should("be.visible");
+      selectInput().realPress("Tab");
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it("should close the list by clicking out of the component", () => {
       CypressMountWithProviders(<MultiSelectComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       body().click({ force: true });
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.be.visible");
+      selectListWrapper().should("not.be.visible");
     });
 
     it.each([
-      ["open", "downarrow"],
-      ["open", "uparrow"],
+      ["open", "ArrowDown"],
+      ["open", "ArrowUp"],
       ["open", "Home"],
       ["open", "End"],
       ["not open", "Enter"],
@@ -807,11 +825,11 @@ context("Tests for Multi Select component", () => {
         CypressMountWithProviders(<MultiSelectComponent />);
 
         commonDataElementInputPreview().focus();
-        selectInput().trigger("keydown", { ...keyCode(key), force: true });
+        selectInput().realPress(key);
         if (state === "open") {
-          selectList().should("be.visible");
+          selectListWrapper().should("be.visible");
         } else {
-          selectList().should("not.be.visible");
+          selectListWrapper().should("not.be.visible");
         }
       }
     );
@@ -824,7 +842,7 @@ context("Tests for Multi Select component", () => {
         dropdownButton().click();
         selectListText(option).click();
         selectInput().should("have.attr", "aria-expanded", "true");
-        selectList().should("be.visible");
+        selectListWrapper().should("be.visible");
         multiSelectPill().should("have.attr", "title", option);
       }
     );
@@ -835,7 +853,7 @@ context("Tests for Multi Select component", () => {
       dropdownButton().click();
       selectListText(option1).click();
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       multiSelectPill().should("have.attr", "title", option1);
       selectListText(option2).click();
       multiSelectPillByPosition(0).should("have.attr", "title", option1);
@@ -850,7 +868,7 @@ context("Tests for Multi Select component", () => {
       dropdownButton().click();
       selectListText(option1).click();
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       multiSelectPill().should("have.attr", "title", option1);
       selectListText(option2).click();
       selectListText(option3).click();
@@ -869,7 +887,7 @@ context("Tests for Multi Select component", () => {
 
         commonDataElementInputPreview().type(text);
         selectInput().should("have.attr", "aria-expanded", "true");
-        selectList().should("be.visible");
+        selectListWrapper().should("be.visible");
         selectOption(positionOfElement("first"))
           .should("have.text", optionValue1)
           .and("be.visible")
@@ -889,7 +907,7 @@ context("Tests for Multi Select component", () => {
       CypressMountWithProviders(<MultiSelectWithLazyLoadingComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       for (let i = 0; i < 3; i++) {
         loader(i).should("be.visible");
       }
@@ -901,7 +919,7 @@ context("Tests for Multi Select component", () => {
       const option = "Amber";
 
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       for (let i = 0; i < 3; i++) {
         loader(i).should("be.visible");
       }
@@ -909,7 +927,7 @@ context("Tests for Multi Select component", () => {
       dropdownButton().click();
       selectResetButton().click({ force: true });
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       for (let i = 0; i < 3; i++) {
         loader(i).should("be.visible");
       }
@@ -925,9 +943,9 @@ context("Tests for Multi Select component", () => {
       selectListText(option).click();
       multiSelectPill().should("have.attr", "title", option);
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       dropdownButton().click().click();
-      selectList()
+      selectListWrapper()
         .find("li")
         .should(($lis) => {
           expect($lis).to.have.length(count);
@@ -939,7 +957,7 @@ context("Tests for Multi Select component", () => {
 
       commonDataElementInputPreview().focus();
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
     });
 
     it("should check list is open when input is clicked and openOnFocus is set", () => {
@@ -947,11 +965,10 @@ context("Tests for Multi Select component", () => {
 
       commonDataElementInputPreview().click();
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
     });
 
-    // FE-5332 logged for bug in master preventing editing of pills when object set as value
-    it.skip("should open correct list and select one when an object is already set as a value", () => {
+    it("should open correct list and select one when an object is already set as a value", () => {
       CypressMountWithProviders(<MultiSelectObjectAsValueComponent />);
 
       multiSelectPill().should("have.attr", "title", option1);
@@ -1047,7 +1064,7 @@ context("Tests for Multi Select component", () => {
       CypressMountWithProviders(<MultiSelectMultiColumnsComponent />);
 
       dropdownButton().click();
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       multiColumnsSelectListHeader()
         .should("have.length", columns)
         .and("be.visible");
@@ -1100,7 +1117,7 @@ context("Tests for Multi Select component", () => {
 
         commonDataElementInputPreview().click().should("be.focused");
         commonDataElementInputPreview().type(text);
-        selectList().should("be.visible");
+        selectListWrapper().should("be.visible");
         multiColumnsSelectListHeader()
           .should("have.length", columns)
           .and("be.visible");
@@ -1183,7 +1200,7 @@ context("Tests for Multi Select component", () => {
       );
 
       multiSelectPill().should("have.attr", "title", "White");
-      commonDataElementInputPreview().type("{backspace}");
+      commonDataElementInputPreview().focus().realPress("Backspace");
       multiSelectPill().should("not.exist");
     });
 
@@ -1194,13 +1211,13 @@ context("Tests for Multi Select component", () => {
       selectListText(option1).click();
       selectListText(option2).click();
       selectInput().should("have.attr", "aria-expanded", "true");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       commonDataElementInputPreview().type("{backspace}");
       multiSelectPillByText(option2).should("not.exist");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
       commonDataElementInputPreview().type("{backspace}");
       multiSelectPill().should("not.exist");
-      selectList().should("be.visible");
+      selectListWrapper().should("be.visible");
     });
 
     it("should have correct hover state of list option", () => {
@@ -1264,7 +1281,7 @@ context("Tests for Multi Select component", () => {
         });
     });
 
-    it("should call onChange event when a list option is selected", () => {
+    it("should call onChange event once when a list option is selected", () => {
       CypressMountWithProviders(<MultiSelectComponent onChange={callback} />);
 
       const position = "first";
@@ -1275,13 +1292,13 @@ context("Tests for Multi Select component", () => {
         .click()
         .then(() => {
           // eslint-disable-next-line no-unused-expressions
-          expect(callback).to.have.been.calledWith({
+          expect(callback).to.have.been.calledOnceWith({
             target: { value: option },
           });
         });
     });
 
-    it.each([["downarrow"], ["uparrow"]])(
+    it.each(["ArrowLeft", "ArrowRight", "ArrowUp", "ArrowDown", "Enter"])(
       "should call onKeyDown event when %s key is pressed",
       (key) => {
         CypressMountWithProviders(
@@ -1290,7 +1307,7 @@ context("Tests for Multi Select component", () => {
 
         commonDataElementInputPreview()
           .focus()
-          .trigger("keydown", { ...keyCode(key), force: true })
+          .realPress(key)
           .then(() => {
             // eslint-disable-next-line no-unused-expressions
             expect(callback).to.have.been.calledOnce;
@@ -1313,6 +1330,45 @@ context("Tests for Multi Select component", () => {
           expect(callback).to.have.been.calledTwice;
           expect(callback.getCalls()[1].args[0]).to.equals(text);
         });
+    });
+  });
+
+  describe("check virtual scrolling", () => {
+    it("renders only an appropriate number of options into the DOM when first opened", () => {
+      CypressMountWithProviders(
+        <MultiSelectWithManyOptionsAndVirtualScrolling />
+      );
+
+      dropdownButton().click();
+
+      selectOptionByText("Option 1.").should("be.visible");
+      selectOptionByText("Option 10.").should("exist").and("not.be.visible");
+      selectOptionByText("Option 30.").should("not.exist");
+    });
+
+    it("changes the rendered options when you scroll down", () => {
+      CypressMountWithProviders(
+        <MultiSelectWithManyOptionsAndVirtualScrolling />
+      );
+
+      dropdownButton().click();
+      selectListWrapper().scrollTo(0, 750).wait(250);
+
+      selectOptionByText("Option 1.").should("not.exist");
+      selectOptionByText("Option 20.").should("be.visible");
+      selectOptionByText("Option 30.").should("exist").and("not.be.visible");
+      selectOptionByText("Option 40.").should("not.exist");
+    });
+
+    it("should filter options when text is typed, taking into account non-rendered options", () => {
+      CypressMountWithProviders(
+        <MultiSelectWithManyOptionsAndVirtualScrolling />
+      );
+
+      commonDataElementInputPreview().type("Option 100");
+      selectOptionByText("Option 100.").should("be.visible");
+      selectOptionByText("Option 1000.").should("be.visible");
+      selectOptionByText("Option 1002.").should("be.visible");
     });
   });
 });